requirements.txt
absl-py==2.1.0
anyio==4.3.0
appdirs==1.4.4
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
docopt==0.6.2
Events==0.5
farm-haystack==1.25.0
filelock==3.13.1
flatbuffers==24.3.6
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
inflect==7.0.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
lazy-imports==0.3.1
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
num2words==0.5.13
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==24.0
platformdirs==4.2.0
posthog==3.5.0
prompthub-py==4.0.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pymongo==4.6.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.14.1